Rename directory from src/main/java to src/main/froggy Update package declerations
