+++ src/main/froggy/AddDeadlineCommand.java
@@ -1,3 +1,5 @@
+package froggy;
+
 public class AddDeadlineCommand extends Command {
 
     private String input;
