+++ src/main/froggy/AddTodoCommand.java
@@ -1,3 +1,5 @@
+package froggy;
+
 public class AddTodoCommand extends Command {
 
     private String input;
