+++ src/main/froggy/Deadline.java
@@ -1,3 +1,5 @@
+package froggy;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
