+++ src/main/froggy/Event.java
@@ -1,3 +1,5 @@
+package froggy;
+
 public class Event extends Task {
 
     protected String from;
