+++ src/main/froggy/FroggyException.java
@@ -1,3 +1,5 @@
+package froggy;
+
 public class FroggyException extends Exception{
     public FroggyException(String message) {
         super(message);
