+++ src/main/froggy/MarkCommand.java
@@ -1,22 +1,24 @@
-public class UnmarkCommand extends Command{
+package froggy;
+
+public class MarkCommand extends Command{
 
     private String input;
 
-    public UnmarkCommand(String input) {
+    public MarkCommand(String input) {
         this.input = input;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
-        if (input.length() <= 7) {
-            System.out.println("Error: Please enter the index of the task to be unmarked.");
+        if (input.length() <= 5) {
+            System.out.println("Error: Please enter the index of the task to be marked.");
             ui.showLine();
         } else {
             try {
-                int index = Integer.parseInt(input.substring(7)) - 1;
+                int index = Integer.parseInt(input.substring(5)) - 1;
                 if (index >= 0 && index < taskList.getSize()) {
-                    taskList.setDone(index, false);
-                    System.out.println("Marked the following task as undone:");
+                    taskList.setDone(index, true);
+                    System.out.println("Marked the following task as done:");
                     taskList.printTask(index);
                     ui.showLine();
                 } else {
@@ -24,7 +26,7 @@ public class UnmarkCommand extends Command{
                     ui.showLine();
                 }
             } catch (NumberFormatException e) {
-                System.out.println("Error: Please enter a valid index after 'unmark'.");
+                System.out.println("Error: Please enter a valid index after 'mark'.");
                 ui.showLine();
             }
         }
