+++ src/main/froggy/Storage.java
@@ -1,3 +1,5 @@
+package froggy;
+
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
