+++ src/main/froggy/Task.java
@@ -1,3 +1,5 @@
+package froggy;
+
 public class Task {
     protected String description;
     protected boolean isDone;
